Git-SDK improvement:
Given OneDrive with the Desktop Redirection feature, and other supported methods to NOT have your Desktop in %USERPROFILE%\Desktop, we should use a more sophisticated method to get the path to the Desktop in order to put a shortcut there. Since we are using Bash we need to step out and invoke a vbs script to get the SpecialFolder called Desktop, and use that as a path.